+++ examples/test-site/gatsby-node.js
@@ -7,6 +7,7 @@
 const TsconfigPathsPlugin = require('tsconfig-paths-webpack-plugin');
 const path = require('path');
 const fs = require('fs');
+const glob = require('glob');
 
 // Fix sourcemap issue
 // See: https://github.com/gatsbyjs/gatsby/issues/6278#issuecomment-402540404
@@ -29,6 +30,17 @@ exports.onCreateWebpackConfig = ({ stage, actions }) => {
         plugins: [new TsconfigPathsPlugin()],
         alias: reactAlias,
       },
+      // On development, we want changes on Bodiless packages to trigger
+      // new builds. Webpack won't watch packages inside node_modules by
+      // default, so we remove the @bodiless folder from its default list.
+      //
+      // See: https://webpack.js.org/configuration/other-options/#snapshot
+      snapshot: {
+        managedPaths: glob.sync(
+          './node_modules/!(@bodiless)*', 
+          { absolute: true },
+        ),
+      }
     });
   }
 };