+++ examples/test-site/src/components/SocialShare/token.tsx
@@ -17,9 +17,11 @@ import {
   withDesign,
   Img,
   Div,
-  Span,
   Label,
+  asToken,
+  removeClasses,
 } from '@bodiless/fclasses';
+import ShareIcon from './icons/Share';
 
 const providersDesign = {
   ProvidersWrapper: addClasses('m-2'),
@@ -38,8 +40,8 @@ const socialShareOrangeDesign = {
   SocialShareProdviders: withDesign(providersDesign),
 };
 
-export const StyledIcon = addClasses('material-icons cursor-pointer align-middle text-white mr-2')(Span);
-export const WhiteIcon = addClasses('material-icons cursor-pointer align-middle text-white')(Span);
+export const StyledIcon = addClasses('fill-current inline-block mr-2')(ShareIcon);
+export const WhiteStyledIcon = asToken(addClasses('text-white'), removeClasses('mr-2'))(StyledIcon);
 export const StyledLabel = addClasses('cursor-pointer')(Label);
 export const Logo = addClasses('bg-blue-500 w-full rounded-full cursor-pointer transition duration-500 ease-in-out hover:bg-black transform hover:-translate-y-1 hover:scale-110')(Img);
 export const LogoNoBackground = addClasses('w-full cursor-pointer')(Img);