+++ packages/gatsby-theme-bodiless/gatsby-config.js
@@ -70,7 +70,7 @@ const plugins = [
 /**
  * Google Fonts plugin.
  */
-if (process.env.GOOGLE_FONTS_ENABLED !== '0') {
+if (process.env.NODE_ENV === 'development' || process.env.GOOGLE_FONTS_ENABLED === '1') {
   plugins.push({
     resolve: 'gatsby-plugin-google-fonts',
     options: {
@@ -162,6 +162,8 @@ if (process.env.BODILESS_DEFAULT_CONTENT_AUTO_DISCOVERY === '1') {
     ),
   );
 }
+
+plugins.push('gatsby-plugin-meta-redirect');
 
 module.exports = {
   siteMetadata: {