+++ packages/gatsby-theme-bodiless/gatsby-node.js
@@ -20,11 +20,10 @@
  */
 const pathUtil = require('path');
 const fs = require('fs');
-
 const { getDisabledPages } = require('@bodiless/components/node-api');
 const { createFilePath } = require('gatsby-source-filesystem');
 const { onCreateNode, createSlug } = require('./create-node');
-
+const createRedirectAlias = require('./create-redirect-alias');
 const Logger = require('./Logger');
 
 const logger = new Logger('gatsby');
@@ -48,6 +47,9 @@ exports.onCreateBabelConfig = args => {
   setBabelPlugin({
     name: '@babel/plugin-proposal-class-properties',
     options: { loose: true },
+  });
+  setBabelPlugin({
+    name: 'babel-plugin-preval',
   });
 };
 
@@ -237,6 +239,7 @@ const createPreviewPagesForTemplates = async ({ actions, graphql, getNode }) => 
 };
 
 exports.createPages = async ({ actions, graphql, getNode }) => {
+  await createRedirectAlias({ actions }, logger);
   await createPagesFromFS({ actions, graphql, getNode });
   if (process.env.NODE_ENV === 'development') {
     await createPreviewPagesForTemplates({ actions, graphql, getNode });